geolocate: Return Unprocessable Content on upstream errors
Previously we were converting upstream errors to 500 Internal Server Error. But
the intention was to return Unprocessable Content (422) when the upstream
errored.

Add testing to the weather handler to make sure this works. This found a bug in
the `text/plain` template: we weren't referencing the emoji properly. Fixed.

Slide in a bit of refactoring while we're here: replace all (hopefully) numbered
HTTP status codes with constants from `http-status-codes`.