Round lat/lon to 2dp
We don't need to issue new queries for every nanometre on the planet. Having a
resolution of a few metres is just fine.

This will reduce queries sent to our providers, and it'll improve the cache hit
ratio a bit too.

Modify the tests to exercise this too, by returning full precision coordinates
and expecting rounded ones.