Fix "unknown" redirects
The way we were trying to handle redirects for the "unknown" cases wasn't right
when there was no trailing slash. Take this example: when the path is `/foo`,
redirecting to `./bar` takes you to `/bar` instead of `/foo/bar`.

Instead we drop relative redirects and figure out the path ourselves.