geocode: Throw "no path" error when location is empty
Previously we were just checking for `undefined`, but we should also check for
empty strings because that's what API Gateway sends when the route is matched,
e.g. for `/metno/` we get an empty `location` path paremeter matching the route
`/metno/{location}`.

Also improve the message slightly to say what to do instead of basically "there
was an error".