Add a feels like temperature to the output
We do this via a liquid.js filter. This lets us write

```
{{ temperature | feels_like: humidity, windSpeed }}
```

in our templates.

A small complication results in the output. We show this as

```
10.0(10.2)°C
```

where the first number is the actual temperature and the second number is the
feels like temperature. But if the feels like is the same as the actual
temperature, then we omit the second number. This means we can't just use
`toString` on a temperature object any more. (Feels like can't just be a
property either, since it requires humidity and wind speed to calculate.)

So now we give the temperature to the rest of the application after rounding,
always, and it can decide when to add the units. This might be revisted later to
make it be more ergonomic. Or it might be fine - we'll see.