metno: Round pressure/humidity/cloud cover to two decimal places
We average these when showing daily summaries, which means we can end up with a
lot of decimal places. 2DP should be plenty.

Since we do rounding in a few places, start a `util` library so we don't end up
writing the same code many times and put the rounding functions in there.