Group hourly weather by day
We're going to want to group entries by hour and then day. This was annoying
before when we had all the hours in a flat array. Update the `Weather` class to
group hours by day. We use `d3-array`'s `InternMap` class to allow us to have
`Date` objects as the keys of our map. This makes the task a bit easier since we
don't have to go back and forth between `Date` objects and strings. And since
we're using `d3-array` we start using its `group` function to simplify doing the
grouping too.

Update the templates to show the hourly data grouped by day now, with a heading
for the date and then the hours underneath that. Still looks terrible, but
something we can build on.