layout: Lay days out on one row, make colour names lowercase
The beginnings of a proper layout. We place the forecasts on one row per day
using CSS grid.

Not all colours were working, because we were using e.g. `lightBlue` instead of
`lightblue`. Fix that by using lowercase consistently.

Most of the CSS file was not used, so trim that down too.