serverless-offline: Only bind to local IPv4 address
Some change in the devcontainer has caused serverless-offline to start binding
to `::1` when no address is specified. Looks like VS Code's auto port forwarding
stuff doesn't yet work in that case, so keep on using v4.

In the geolocator we have a hardcoded result for local testing (Greenwich). This
works by looking at the IP. That's easy to fix for IPv6-localhost (`::1`), so we
can do that.