static: Fix last-modified headers, move index handler to static
`serverless-esbuild`, which we use to build the project, doesn't preserve file
modification times. We make use of these times to generate the `last-modified`
headers for static files, which we bundle in the Lambda `.zip`s.  I [submitted a
PR upstream][pr] to fix this. Switch to using a fork containing this fix.

The `index` handler was implementing very similar logic to the `static` handler,
but slightly differently. We move the `index` handler to the `static` module and
make it use the same `fileData` as the `static` handler. This allows us to
remove the `index` handler entirely.

Add tests for both handlers. This uses a test file stored in the repo. But note
that `git` doesn't preserve file modification times, so we have to work these
out dynamically in the tests.

Fix a nit too: handlers don't have to be async. They can simply return their
result directly, no need to wrap it in a `Promise.resolve`.

[pr]: https://github.com/floydspace/serverless-esbuild/pull/539