util: Utility functions for adding/removing prefixes and suffixes
Add `addPrefix` and `addSuffix` functions to the `util` module and use them in
the few places we were implementing this functionality manually.

Using the techniques described in [this article by Javier
Casas][dependent-types], we can have the functions take `string | undefined` and
return the same type as the input - a return type depending on an input value,
i.e. dependent types. In practice this lets us use the functions in a
non-annoying manner: if the input string is type `string` (no `| undefined`),
then we don't have to deal with the `undefined` case since the return type will
be `string`. If it's `string | undefined`, then we get `string | undefined` as
the return type, which is what we want.

[dependent-types]: https://www.javiercasas.com/articles/typescript-dependent-types