static: Use the cache busting pattern to serve static assets
We had a problem previously. Taking the `style.css` file as an example: since
the file is served at `/style.css` and with `cache-control` headers set to
`public, s-maxage=60`, browsers and caches will cache it. Even though we set an
ETag and a `Last-Modified` header, the browser will not check for changes until
the cache expires. This means that if we update the file and the HTML
accordingly, users could be using the old CSS for a while, and therefore see a
broken layout.

There's a pattern called "cache busting" that solves this problem. The idea is
to change the URL of the file every time it changes. This way, the browser will
be forced to download the new file. We do this by using a scheme like

```
/static/style.sha256@<hash>.css
```

for the URL of the file. The `<hash>` part is the SHA256 hash of the file, which
is the same as is in the ETag.

Implement this in the `static` handler by having it handle URLs in this format,
serving the underlying file if it exists and the hash matches. If the hash does
not match, the handler returns a 404.

Update the Liquid templates accordingly, to generate URLs with hashes. For that
we needed to move `fileInfo` to a library module, so that the static handler and
the weather handler can both use it, since they both need to know the hashes of
the static assets.