ci: Read node version from package.json, set node version in package.json
The version of node we were using in the devcontainer and in CI were different
(22 vs. 20). If we put an "engines" field in `package.json` then at least users
will be notified when this gets out of sync, and we can have CI read this from
the file instead of having the value itself. We could consider adding a `.npmrc`
with `engine-strict` set to `true` to make this an error - undecided right now.

Also drop `cache-dependency-file`, since `actions/setup-node` will find
`yarn.lock` by default anyway.