Move test config to workspace
This moves the Jest configuration from the root `package.json` to the
`coldoutsi.de` workspace's `package.json`.

It's a leftover from an attempt to use Cloudflare Workers, and have their
sources in another workspace. We've abandoned that approach, but this is
still useful in csae we want to introduce more workspaces in the future.