Migrate to pnpm
There's some obscure errors about `ts-node` occurring. I tried to debug this for
a few hours [over on the `rssfilter` proejct][rssfilter-pr] but didn't really
get anywhere. Switching to `pnpm` fixed it there, and it seems to here too.

This has brought a bump of `middy` some other deps which required minor changes.

[rssfilter-pr]: https://github.com/iainlane/rssfilter/pull/98