geolocate: Reduce duplication in middleware tests
There were a lot of repeated patterns in the tests. This change removes most of
that.

Additionally, we now always convert the path parameters (lat and lon) to an
empty string if either is missing. Since we're behind an API Gateway in
production, we should be able to expect that this is the case for real requests
--- therefore, it's better if the tests behave like this too and we don't handle
`undefined` there.